Move the new voluntary conversion project
We had this at a temporary path used to test the new task list. Now we
want to move it to it's final home at:

`/conversion/voluntary/projects/new`

Worth noting that Rails resourceful routes put the POST path at
`conversion_voluntary_projects` not `new_conversion_voluntary_projects`.
