+++ app/views/projects/shared/_new_projects.html.erb
@@ -1,7 +1,7 @@
 <% if policy(:project).new? %>
   <div class="govuk-button-group">
     <%= link_to t("conversion_project.voluntary.new.title"),
-          conversion_voluntary_new_path,
+          new_conversions_voluntary_project_path,
           class: "govuk-button",
           data: {module: "govuk-button"} %>
     <%= link_to t("conversion_project.involuntary.new.title"),
