+++ spec/accessibility/projects_spec.rb
@@ -31,7 +31,7 @@ RSpec.feature "Test projects accessibility", driver: :headless_firefox, accessib
   end
 
   scenario "new voluntary conversion projects page" do
-    visit conversion_voluntary_new_path
+    visit new_conversions_voluntary_project_path
 
     expect(page).to have_content(I18n.t("conversion_project.voluntary.new.title"))
     expect(page).to be_axe_clean
