+++ spec/features/regional_delivery_officers_can_create_new_projects_spec.rb
@@ -10,7 +10,7 @@ RSpec.feature "Regional delivery officers can create new projects" do
       visit projects_path
 
       expect(page)
-        .to have_link I18n.t("conversion_project.voluntary.new.title"), href: conversion_voluntary_new_path
+        .to have_link I18n.t("conversion_project.voluntary.new.title"), href: new_conversions_voluntary_project_path
     end
 
     it "shows a button that adds a new involuntary conversion" do
