+++ spec/requests/project_management_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe "Project management" do
     context "when the user is a regional delivery officer" do
       let(:user) { create(:user, :regional_delivery_officer) }
       it "allows the action and renders the new template" do
-        get conversion_voluntary_new_path
+        get new_conversions_voluntary_project_path
         expect(response).to have_http_status(:success)
         expect(response).to render_template(:new)
       end
@@ -24,7 +24,7 @@ RSpec.describe "Project management" do
       let(:user) { create(:user, :caseworker) }
 
       it "redirects to the root path and displays a not authorized message" do
-        get conversion_voluntary_new_path
+        get new_conversions_voluntary_project_path
         expect(response).not_to render_template(:new)
         expect(response).to redirect_to(root_path)
         follow_redirect!
@@ -41,7 +41,7 @@ RSpec.describe "Project management" do
       let(:user) { create(:user, :team_leader) }
 
       it "redirects to the root path and displays a not authorized message" do
-        get conversion_voluntary_new_path
+        get new_conversions_voluntary_project_path
         expect(response).not_to render_template(:new)
         expect(response).to redirect_to(root_path)
         follow_redirect!
